Feature/multi file database (#23)
* multi-file database init commit

* multi-file database first sketch

* add methods body

* working saveInvoices and getInvoices

* working findInvoiceById

* multi-file database init commit

* multi-file database first sketch

* add methods body

* working saveInvoices and getInvoices

* working findInvoiceById

* creates tests

* Add cache

* Add cache and tests

* Improve logic, add abstraction db class, od mapper config provider

* add properties to tests

* correct some errors

* separate reading and writing to idFile

* add Spring dependency to Cache and Mapper

* correct MapperProvider, add different file path for tests

* revert conditions

+++ src/main/java/pl/futurecollars/invoices/service/InvoiceController.java
@@ -56,6 +56,9 @@ public class InvoiceController implements InvoiceApi {
     }
 
     public void deleteInvoice(@PathVariable("id") long id) {
+        if (invoiceService.getInvoice(id).isEmpty()) {
+            throw new InvoiceNotFoundException(id);
+        }
         invoiceService.deleteInvoice(id);
     }
 }
